v2/requirements.txt
asgiref==3.2.10
boto==2.49.0
boto3==1.14.16
botocore==1.17.16
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Cython==0.29.14
Django==3.0.8
docutils==0.15.2
Flask==1.1.2
gensim==3.8.3
idna==2.10
image==1.5.32
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
nltk==3.5
numpy==1.19.0
pandas==1.0.5
Pillow==7.2.0
pyparsing==2.4.7
pypi==2.1
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
requests==2.24.0
s3transfer==0.3.3
scikit-learn==0.23.1
scipy==1.5.0
six==1.15.0
sklearn==0.0
smart-open==2.1.0
sqlparse==0.3.1
textblob==0.15.3
threadpoolctl==2.1.0
tqdm==4.47.0
urllib3==1.25.9
Werkzeug==1.0.1
